# openapi.yaml
...
openapi: 3.1.0
info:
  title: Liquid Asset Registry
  summary: API for registering, updating, deregistering, querying, and auditing Liquid asset metadata.
  description: FastAPI implementation of the Liquid Asset Registry v2 service with legacy v1 compatibility.
  version: 0.1.0
paths:
...
  /:
    get:
      tags:
      - Legacy
      summary: Legacy all-assets endpoint
      description: Return the legacy-compatible asset object keyed by asset ID.
      operationId: getAllAssetsLegacyRoot
      responses:
        '200':
          description: Legacy asset listing keyed by asset ID.
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Malformed or semantically invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication or signature verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated actor is not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource or route was not found.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - $ref: '#/components/schemas/FrameworkErrorResponse'
                title: Response 404 Getallassetslegacyroot
        '409':
          description: The request conflicts with current registry state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The request body exceeds the configured size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required verification service is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Legacy
      summary: Legacy registration endpoint
      description: Register an asset through the legacy-compatible flow. New clients should use `POST /v2/assets`.
      operationId: registerAssetLegacyRoot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegacyAssetRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Registerassetlegacyroot
        '400':
          description: Malformed or semantically invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication or signature verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated actor is not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource or route was not found.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - $ref: '#/components/schemas/FrameworkErrorResponse'
                title: Response 404 Registerassetlegacyroot
        '409':
          description: The request conflicts with current registry state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The request body exceeds the configured size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required verification service is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: The request exceeded a registration or domain-proof fetch limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /index.json:
    get:
      tags:
      - Legacy
      summary: Legacy all-assets JSON endpoint
      description: Return the legacy-compatible asset object keyed by asset ID.
      operationId: getAllAssetsLegacyIndex
      responses:
        '200':
          description: Legacy asset listing keyed by asset ID.
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Malformed or semantically invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication or signature verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated actor is not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource or route was not found.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - $ref: '#/components/schemas/FrameworkErrorResponse'
                title: Response 404 Getallassetslegacyindex
        '409':
          description: The request conflicts with current registry state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The request body exceeds the configured size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required verification service is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /icons.json:
    get:
      tags:
      - Legacy
      summary: Get approved asset icons
      description: Return current approved icons as raw Base64 values keyed by asset ID.
      operationId: getIconsLegacy
      responses:
        '200':
          description: Base64 PNG icons keyed by asset ID.
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        '400':
          description: Malformed or semantically invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication or signature verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated actor is not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource or route was not found.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - $ref: '#/components/schemas/FrameworkErrorResponse'
                title: Response 404 Geticonslegacy
        '409':
          description: The request conflicts with current registry state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The request body exceeds the configured size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required verification service is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /contract/validate:
    post:
      tags:
      - Legacy
      summary: Legacy contract validation endpoint
      description: Validate legacy version 0 contract metadata against its supplied contract hash.
      operationId: validateContractLegacy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegacyContractValidationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Malformed or semantically invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication or signature verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated actor is not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource or route was not found.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - $ref: '#/components/schemas/FrameworkErrorResponse'
                title: Response 404 Validatecontractlegacy
        '409':
          description: The request conflicts with current registry state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The request body exceeds the configured size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required verification service is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /{asset_id}:
    get:
      tags:
      - Legacy
      summary: Legacy asset lookup endpoint
      description: Return one active asset in the legacy-compatible response shape.
      operationId: getAssetLegacyRoot
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          minLength: 64
          maxLength: 64
          pattern: ^[0-9a-fA-F]{64}$
          title: Asset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Getassetlegacyroot
        '400':
          description: Malformed or semantically invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication or signature verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated actor is not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource or route was not found.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - $ref: '#/components/schemas/FrameworkErrorResponse'
                title: Response 404 Getassetlegacyroot
        '409':
          description: The request conflicts with current registry state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The request body exceeds the configured size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required verification service is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Legacy
      summary: Legacy asset deregistration endpoint
      description: Deregister an active legacy asset after verifying its legacy deletion signature.
      operationId: deleteAssetLegacyRoot
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          minLength: 64
          maxLength: 64
          pattern: ^[0-9a-fA-F]{64}$
          title: Asset Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegacyDeletionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Malformed or semantically invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication or signature verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated actor is not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource or route was not found.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - $ref: '#/components/schemas/FrameworkErrorResponse'
                title: Response 404 Deleteassetlegacyroot
        '409':
          description: The request conflicts with current registry state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The request body exceeds the configured size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required verification service is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v2/assets/{asset_id}:
    get:
      tags:
      - Assets
      summary: Get an asset by asset ID
      description: Return the active v2-normalized asset record for an asset ID.
      operationId: getAssetV2
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          minLength: 64
          maxLength: 64
          pattern: ^[0-9a-fA-F]{64}$
          title: Asset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          description: Malformed or semantically invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication or signature verification failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated actor is not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The requested resource or route was not found.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - $ref: '#/components/schemas/FrameworkErrorResponse'
                title: Response 404 Getassetv2
        '409':
          description: The request conflicts with current registry state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: The request body exceeds the configured size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required verification service is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AssetIconDescriptor:
      properties:
        href:
          type: string
          pattern: ^/v2/assets/[0-9a-f]{64}/icon/[0-9a-f]{64}\.png$
          title: Href
      additionalProperties: false
      type: object
      required:
      - href
      title: AssetIconDescriptor
...
    AssetResponse:
      properties:
        asset_id:
          type: string
          maxLength: 64
          minLength: 64
          pattern: ^[0-9a-fA-F]{64}$
          title: Asset Id
        contract:
          $ref: '#/components/schemas/ContractMetadataResponse'
        initial_issuer_pubkey:
          type: string
          maxLength: 66
          minLength: 66
          pattern: ^(02|03)[0-9a-fA-F]{64}$
          title: Initial Issuer Pubkey
        initial_issuer_pubkey_source:
          type: string
          title: Initial Issuer Pubkey Source
        current_issuer_pubkey:
          type: string
          maxLength: 66
          minLength: 66
          pattern: ^(02|03)[0-9a-fA-F]{64}$
          title: Current Issuer Pubkey
        issuer_pubkey_history:
          items:
            $ref: '#/components/schemas/IssuerPubkeyHistoryEntry'
          type: array
          title: Issuer Pubkey History
        mutable:
          $ref: '#/components/schemas/MutableMetadata-Output'
        admin:
          anyOf:
          - $ref: '#/components/schemas/AdminAnnotations'
          - type: 'null'
        icon:
          anyOf:
          - $ref: '#/components/schemas/AssetIconDescriptor'
          - type: 'null'
        status:
          type: string
          enum:
          - active
          - deregistered
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - asset_id
      - contract
      - initial_issuer_pubkey
      - initial_issuer_pubkey_source
      - current_issuer_pubkey
      - mutable
      - icon
      - status
      - created_at
      - updated_at
      title: AssetResponse
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
      additionalProperties: false
      type: object
      required:
      - error
      title: ErrorResponse
...
    FrameworkErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
      additionalProperties: false
      type: object
      required:
      - detail
      title: FrameworkErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LegacyAssetRequest:
      properties:
        asset_id:
          type: string
          maxLength: 64
          minLength: 64
          pattern: ^[0-9a-fA-F]{64}$
          title: Asset Id
        contract:
          properties:
            version:
              type: integer
              const: 0
              title: Version
            issuer_pubkey:
              type: string
              maxLength: 66
              minLength: 66
              pattern: ^(02|03)[0-9a-fA-F]{64}$
              title: Issuer Pubkey
            name:
              type: string
              maxLength: 255
              minLength: 1
              pattern: ^[\x01-\x7f]+$
              title: Name
            ticker:
              anyOf:
              - type: string
                maxLength: 24
                minLength: 3
                pattern: ^[A-Za-z0-9.\-]+$
              - type: 'null'
              title: Ticker
            collection:
              anyOf:
              - type: string
                maxLength: 255
                minLength: 1
                pattern: ^[\x01-\x7f]+$
              - type: 'null'
              title: Collection
            precision:
              type: integer
              maximum: 8.0
              minimum: 0.0
              title: Precision
              default: 0
            entity:
              $ref: '#/components/schemas/LegacyContractEntity'
          additionalProperties:
            $ref: '#/components/schemas/LegacyExtraValue-Input'
          propertyNames:
            pattern: ^[^\x00]*$
          type: object
          required:
          - version
          - issuer_pubkey
          - name
          - entity
          title: LegacyContractMetadata
          description: Legacy version 0 contract metadata. The canonical serialized contract object is limited to 4096 bytes;
            this default can be changed with ASSET_REGISTRY_LEGACY_CONTRACT_MAX_BYTES. At most 32 arbitrary extra fields are
            accepted, and each serialized extra-field value is limited to 2048 bytes.
        domain_verification_method:
          anyOf:
          - type: string
            enum:
            - dns
            - http
          - type: 'null'
          title: Domain Verification Method
          default: http
      additionalProperties:
        $ref: '#/components/schemas/LegacyExtraValue-Input'
      propertyNames:
        pattern: ^[^\x00]*$
      type: object
      required:
      - asset_id
      - contract
      title: LegacyAssetRequest
    LegacyContractEntity:
      properties:
        domain:
          type: string
          maxLength: 255
          minLength: 3
          pattern: ^(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\.)+[a-z](?:[a-z0-9-]{0,61}[a-z0-9])?\.?$
          title: Domain
      additionalProperties: false
      type: object
      required:
      - domain
      title: LegacyContractEntity
...
    LegacyContractValidationRequest:
      properties:
        contract:
          properties:
            version:
              type: integer
              const: 0
              title: Version
            issuer_pubkey:
              type: string
              maxLength: 66
              minLength: 66
              pattern: ^(02|03)[0-9a-fA-F]{64}$
              title: Issuer Pubkey
            name:
              type: string
              maxLength: 255
              minLength: 1
              pattern: ^[\x01-\x7f]+$
              title: Name
            ticker:
              anyOf:
              - type: string
                maxLength: 24
                minLength: 3
                pattern: ^[A-Za-z0-9.\-]+$
              - type: 'null'
              title: Ticker
            collection:
              anyOf:
              - type: string
                maxLength: 255
                minLength: 1
                pattern: ^[\x01-\x7f]+$
              - type: 'null'
              title: Collection
            precision:
              type: integer
              maximum: 8.0
              minimum: 0.0
              title: Precision
              default: 0
            entity:
              $ref: '#/components/schemas/LegacyContractEntity'
          additionalProperties:
            $ref: '#/components/schemas/LegacyExtraValue-Input'
          propertyNames:
            pattern: ^[^\x00]*$
          type: object
          required:
          - version
          - issuer_pubkey
          - name
          - entity
          title: LegacyContractMetadata
          description: Legacy version 0 contract metadata to validate. The canonical serialized contract object is limited
            to 4096 bytes by default; this can be changed with ASSET_REGISTRY_LEGACY_CONTRACT_MAX_BYTES. At most 32 arbitrary
            extra fields are accepted, and each serialized extra-field value is limited to 2048 bytes.
        contract_hash:
          type: string
          maxLength: 64
          minLength: 64
          pattern: ^[0-9a-fA-F]{64}$
          title: Contract Hash
      additionalProperties: false
      type: object
      required:
      - contract
      - contract_hash
      title: LegacyContractValidationRequest
    LegacyDeletionRequest:
      properties:
        signature:
          type: string
          minLength: 1
          title: Signature
      additionalProperties: false
      type: object
      required:
      - signature
      title: LegacyDeletionRequest
...